simplify logic
